+++ src/config/cli.rs
@@ -1,9 +1,16 @@
 use clap::Arg;
 
-pub fn account_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("account")
-        .long("account")
-        .short("a")
-        .help("Selects a specific account")
-        .value_name("STRING")
+pub fn config_args<'a>() -> Vec<Arg<'a, 'a>> {
+    vec![
+        Arg::with_name("config")
+            .long("config")
+            .short("c")
+            .help("Forces a specific config path")
+            .value_name("PATH"),
+        Arg::with_name("account")
+            .long("account")
+            .short("a")
+            .help("Selects a specific account")
+            .value_name("NAME"),
+    ]
 }